noteshub/backend/app/utils/auth.py
from datetime import datetime, timedelta
from typing import Optional
import bcrypt
from jose import JWTError, jwt
from passlib.context import CryptContext
from fastapi import Depends, HTTPException, status
from fastapi.security import HTTPBearer, HTTPAuthorizationCredentials
from bson import ObjectId
from bson.errors import InvalidId
from app.config import settings
from app.database import get_database
from app.schemas.user import TokenData

# New passwords use pbkdf2_sha256. Legacy bcrypt hashes are verified via bcrypt.checkpw (avoids passlib/bcrypt backend quirks).
pwd_context = CryptContext(schemes=["pbkdf2_sha256"], deprecated="auto")

# HTTP Bearer token
security = HTTPBearer()


def verify_password(plain_password: str, hashed_password: str) -> bool:
    """Verify a password against its hash"""
    try:
        if pwd_context.verify(plain_password, hashed_password):
            return True
    except (ValueError, TypeError):
        pass
    h = hashed_password or ""
    if h.startswith("$2a$") or h.startswith("$2b$") or h.startswith("$2y$"):
        try:
            return bcrypt.checkpw(
                plain_password.encode("utf-8"),
                hashed_password.encode("utf-8"),
            )
        except (ValueError, TypeError):
            return False
    return False


def get_password_hash(password: str) -> str:
    """Hash a password"""
    return pwd_context.hash(password)


def create_access_token(data: dict, expires_delta: Optional[timedelta] = None) -> str:
    """Create a JWT access token"""
    to_encode = data.copy()
    if expires_delta:
        expire = datetime.utcnow() + expires_delta
    else:
        expire = datetime.utcnow() + timedelta(minutes=settings.ACCESS_TOKEN_EXPIRE_MINUTES)
    
    to_encode.update({"exp": expire})
    encoded_jwt = jwt.encode(to_encode, settings.SECRET_KEY, algorithm=settings.ALGORITHM)
    return encoded_jwt


async def get_current_user(
    credentials: HTTPAuthorizationCredentials = Depends(security)
):
    """Get the current authenticated user"""
    credentials_exception = HTTPException(
        status_code=status.HTTP_401_UNAUTHORIZED,
        detail="Could not validate credentials",
        headers={"WWW-Authenticate": "Bearer"},
    )
    
    try:
        token = credentials.credentials
        payload = jwt.decode(token, settings.SECRET_KEY, algorithms=[settings.ALGORITHM])
        user_id: str = payload.get("sub")
        if user_id is None:
            raise credentials_exception
        token_data = TokenData(user_id=user_id)
        user_object_id = ObjectId(token_data.user_id)
    except (JWTError, InvalidId, ValueError, TypeError):
        raise credentials_exception
    
    db = get_database()
    user = await db.users.find_one({"_id": user_object_id, "deletedAt": None})
    if user is None:
        raise credentials_exception
    
    return user


async def get_current_user_id(current_user: dict = Depends(get_current_user)) -> str:
    """Get the current user's ID as a string"""
    return str(current_user["_id"])
